python/chatbot-service/requirements.txt
# Infrastructure (Cực nhẹ)
aio-pika
minio
redis
python-dotenv
requests
aiohttp
google-genai

# LangChain & Vector
langchain-core
langchain-community
langchain-text-splitters
chromadb-client

# Reranking bản Lite
flashrank
onnxruntime

# File Processing
pypdf
docx2txt
openpyxl

# Utils
tenacity
tqdm
pycryptodome